delegate peer id generation to server
