--- mobile-browser-based-version/helper_servers/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "helper-servers",
-  "version": "0.1.0",
-  "scripts": {
-    "proxy": "node reverse_proxied/proxy_server.js",
-    "peer": "node reverse_proxied/peer_server.js",
-    "task": "node reverse_proxied/task_server.js",
-    "helper": "node single_server/helper_server.js"
-  },
-  "dependencies": {
-    "express": "^4.17.1",
-    "http-proxy": "^1.18.1",
-    "peer": "^0.6.1"
-  }
-}
