+++ mobile-browser-based-version/package.json
@@ -5,7 +5,11 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "lint": "vue-cli-service lint",
+    "proxy": "node src/centralized_server/reverse_proxied/proxy_server.js",
+    "peer": "node src/centralized_server/reverse_proxied/peer_server.js",
+    "task": "node src/centralized_server/reverse_proxied/task_server.js",
+    "helper": "node src/centralized_server/single_server/helper_server.js"
   },
   "dependencies": {
     "@meforma/vue-toaster": "^1.2.2",
@@ -23,7 +27,10 @@
     "vue-chartjs": "^3.5.1",
     "vue-loaders": "^4.1.4",
     "vue-notification-renderless": "^0.7.3",
-    "vue-router": "^4.0.0-0"
+    "vue-router": "^4.0.0-0",
+    "express": "^4.17.1",
+    "http-proxy": "^1.18.1",
+    "peer": "^0.6.1"
   },
   "devDependencies": {
     "@tailwindcss/postcss7-compat": "^2.0.4",
